requirements.txt
solana>=0.32.0,<0.33.0
solders>=0.19.0,<0.20.0
pynacl>=1.5.0,<2.0.0
# anchorpy - if/when Anchor on-chain programs are used and an SDK is generated/used 